# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartFin API",
    "version": "0.1.0",
    "description": "Personal and family finance API. Amounts are decimal strings; instants are ISO 8601 UTC; financial dates are YYYY-MM-DD."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sf_session",
        "description": "Session cookie set by sign-in (named __Host-sf_session over HTTPS). State-changing requests also need the X-SmartFin-CSRF header."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token for the Android app (sign in with X-SmartFin-Client: android)."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "bad_request",
                  "validation_failed",
                  "unauthenticated",
                  "forbidden",
                  "csrf_failed",
                  "not_found",
                  "conflict",
                  "rate_limited",
                  "internal_error",
                  "service_unavailable"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Every API error uses this shape."
      },
...
      "ImportPreview": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rowNumber": {
                  "type": "integer"
                },
                "valueDate": {
                  "type": "string",
                  "format": "date",
                  "description": "Calendar date",
                  "example": "2026-10-04"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "income",
                    "expense"
                  ]
                },
                "amount": {
                  "type": "string",
                  "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
                  "description": "Positive amount",
                  "example": "123456.78"
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "reference": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "duplicate": {
                  "type": "string",
                  "enum": [
                    "none",
                    "possible",
                    "likely"
                  ],
                  "description": "likely: same reference, or same date, amount and description as an existing transaction (or an earlier row). possible: same amount within 2 days."
                },
                "matchTransactionId": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                }
              },
              "required": [
                "rowNumber",
                "valueDate",
                "type",
                "amount",
                "description",
                "reference",
                "duplicate",
                "matchTransactionId"
              ]
            }
          },
          "counts": {
            "type": "object",
            "properties": {
              "total": {
                "type": "number"
              },
              "likely": {
                "type": "number"
              },
              "possible": {
                "type": "number"
              }
            },
            "required": [
              "total",
              "likely",
              "possible"
            ]
          }
        },
        "required": [
          "rows",
          "counts"
        ]
      },
      "ImportPreviewInput": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rowNumber": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "description": "Row number in the source file"
                },
                "valueDate": {
                  "type": "string",
                  "format": "date",
                  "description": "Calendar date",
                  "example": "2026-10-04"
                },
                "amount": {
                  "type": "string",
                  "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
                  "description": "Signed: negative is money out (expense), positive is money in (income).",
                  "example": "123456.78"
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "maxLength": 200
                },
                "reference": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "maxLength": 64
                }
              },
              "required": [
                "rowNumber",
                "valueDate",
                "amount"
              ]
            },
            "minItems": 1,
            "maxItems": 5000
          }
        },
        "required": [
          "accountId",
          "rows"
        ]
      },
      "ImportResult": {
        "type": "object",
        "properties": {
          "batchId": {
            "type": "string",
            "format": "uuid"
          },
          "imported": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          },
          "replayed": {
            "type": "boolean",
            "description": "True when this Idempotency-Key was already used; nothing was imported again."
          }
        },
        "required": [
          "batchId",
          "imported",
          "skipped",
          "replayed"
        ]
      },
      "ImportCommitInput": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rowNumber": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "description": "Row number in the source file"
                },
                "valueDate": {
                  "type": "string",
                  "format": "date",
                  "description": "Calendar date",
                  "example": "2026-10-04"
                },
                "amount": {
                  "type": "string",
                  "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
                  "description": "Signed: negative is money out (expense), positive is money in (income).",
                  "example": "123456.78"
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "maxLength": 200
                },
                "reference": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "maxLength": 64
                },
                "include": {
                  "type": "boolean"
                },
                "categoryId": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                }
              },
              "required": [
                "rowNumber",
                "valueDate",
                "amount",
                "include"
              ]
            },
            "minItems": 1,
            "maxItems": 5000
          }
        },
        "required": [
          "accountId",
          "fileName",
          "rows"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/imports/preview": {
      "post": {
        "tags": [
          "Imports"
        ],
        "summary": "Check statement rows for duplicates before importing",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPreviewInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/commit": {
      "post": {
        "tags": [
          "Imports"
        ],
        "summary": "Import statement rows (idempotent)",
        "description": "Retrying with the same Idempotency-Key returns the first result with replayed: true.",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            },
            "required": true,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCommitInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}